requirements.txt
google-cloud-storage==1.19.1
gunicorn==19.9.0
aniso8601==8.0.0
appdirs==1.4.3
astroid==2.3.1
atomicwrites==1.3.0
attrs==19.2.0
black==19.3b0
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
numpy==1.17.2
packaging==19.2
Pillow==6.2.0
pluggy==0.13.0
py==1.8.0
pylint==2.4.2
pyparsing==2.4.2
pytest==5.2.1
python-dotenv==0.10.3
pytz==2019.3
PyYAML==5.1.2
six==1.12.0
toml==0.10.0
torch==1.3.0
torchvision==0.4.1
typed-ast==1.4.0
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
